back_end/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic>=2.6.0
pydantic[email]
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
chromadb>=0.5.0
requests==2.31.0
email-validator==2.1.0
openai>=1.0.0
Pillow>=10.0.0
google.genai`>=0.1.0